requirements.txt
# ============================================
# REQUIREMENTS - Crop Recommendation System
# ============================================

# Core ML Libraries
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.2

# Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Werkzeug>=3.0.1

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration
PyYAML>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
pylint>=2.17.5

# Deployment
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.23

# Optional - Enhanced Features
# mlflow==2.8.0
# fastapi==0.104.1
# uvicorn==0.24.0
# streamlit==1.28.0
